+++ src/source/uvcoord_font_si.h
@@ -21,7 +21,7 @@
 
 
 
-extern const unsigned int font_si_texcoords[] = {
+extern const uint16_t font_si_texcoords[] = {
  0, 0, 8, 9,				// 0
  8, 0, 8, 9,				// 1
  16, 0, 8, 9,				// 2
